requirements.txt
pyyaml
python_version=='3.9.6'